requirements.txt
numpy
pandas
python-dotenv
pymongo
sentence-transformers
torch 
pydantic
langchain
langchain-google-genai
langchain-community
